CrOS - Change file browser to use new ExtensionDialog window.
The CrOS file browser was initially built to use HtmlDialogView, which does not really understand extensions.  This CL implements an ExtensionDialog similar to an ExtensionPopup, and uses it for the file browser.  This allows us to return to using window.close() in the file manager extension JavaScript and hopefully will make the file browser less brittle going forward.

BUG=chromium-os:15415
TEST=Hit control-O to open files.  Try opening a file, canceling with the cancel button, and canceling with escape.  Do the same with "save as" for images.  Automated tests to follow.

Review URL: http://codereview.chromium.org/7034046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@86099 0039d316-1c4b-4281-b951-d872f2087c98