Revert "Revert 84829 - Initial implementation of "Synced Bookmarks" folder."
Second attempt to land this change. Fixes memory leak (82186), 
deleted bookmarks (82273) and crash on add bookmarks in windows (82349).

Original desscription:
Mostly ensures that sycned bookmarks are correctly treated as an immutable
folder. Some of the UI-bits maybe incomplete. For example, if
enable-synced-bookmarks-folder is set, then synced bookmarks will appear in the
bookmark manager page and some components of the UI but it's not on the bookmark
bar or anything like that. This change also ensures that the synced bookmark
folder matches a sync folder if one is available.


BUG=
TEST=test bookmark addition/moving around in combination with sync


Review URL: http://codereview.chromium.org/7012005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@86902 0039d316-1c4b-4281-b951-d872f2087c98