mac: turn download shelf link into a button
xib changes:
* Change button cell from HyperlinkButtonCell to DownloadShowAllCell
* Disconnect |showAllCell_|
* Change button text from IDS_SHOW_ALL_DOWNLOADS to IDS_SHOW_ALL
* Make button higher
* Give button class DownloadShowAllButton
* Remove NSImageView as the image is now on the button
* Reduce padding between "show all" button and close button by 2

DownloadShowAllCell is a new class that looks like a button with an icon. Some common logic with the download items has been extracted.

BUG=79502
TEST=Download something, look at bottom right corner.
With several downloads, the new button should interact with the download items in a sane way on window resize.
It should look right with themes.

Review URL: http://codereview.chromium.org/7084009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@87232 0039d316-1c4b-4281-b951-d872f2087c98