Apply content-security-policy to chrome://plugins page.  This involves converting all inline javascript to external .js files, since CSP can't determine whether inline script is legitimate or an XSS.
We move the jstemplate_compiled.js file to the share resources directory,
since several WEBUI components will need this, and update the compile.py
script that produces it to write to this new location.

We move the javascript out of plugins.html to plugins.js; note that the
sections at line 130 and line 251 are new.
Review URL: http://codereview.chromium.org/7086003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@87358 0039d316-1c4b-4281-b951-d872f2087c98