Add preedit-text area for ChromeOS IME candidate window.
This patch is the third-step for enabling off-the-spot IME on Pepper on ChromeOS.
When the focus is on a PPAPI plugin that can not draw preedit text by themselves
(since current PPAPI have no support for IMEs, this applies to all PPAPI plugins
for now), Chrome take care of it. To this end, the candidate window need to be
able to draw preedit text on it.

BUG=chromium-os:13374
TEST=1. Please apply http://codereview.chromium.org/7041003 (under review) which switches
on/off preedit display (via gtk_im_context API) depending on focus information.
2. Open a page with PPAPI plugin (e.g., weathernews.jp with PPAPI flash, or the
example plugin in src/ppapi/example)
3. Focus the plugin
(Actually, http://crosbug.com/16500 seems blocking this to work. So for testing,
please focus the plugin, unfocus by clicking outer html area, and re-focus)
4. Turn ime on and type some string.
5. Verify preedit text is displayed in addition to the candidate table.

Proper placement of the candidate window is not included in this patch (it'll
be displayed on a rather random position.)

Review URL: http://codereview.chromium.org/7189008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@89643 0039d316-1c4b-4281-b951-d872f2087c98