ui/base/models: Replace ASSERT_* checks by EXPECT_* in TreeNode unittests.
It turns out that we generally should use EXPECT, and only switch to ASSERT when
the test would otherwise crash.

An example when to use ASSERT is:
- When going to manipulate a returned pointer value, we should ASSERT at some
  point that the pointer is non-NULL.

Another example is:
- If you are going to check the values in an array, you ASSERT on the size of
  the array. I.e, the array should contain 3 elements, because you are going to
  access the third element at index 2, if the array doesn't contain 3
  elements, the test will crash, so you should ASSERT before trying to
  access the element instead of EXPECT because you want to abort the test
instead of crashing the entire binary.

And in the cases here, we don't have these situations and the individual test
expectations are not linked, so we are better using EXPECT than ASSERT
everywhere.

BUG=None
TEST=ui_unittests --gtest_filter=TreeNodeModelTest.*

R=phajdan.jr@chromium.org,sky@chromium.org

Review URL: http://codereview.chromium.org/7243005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@90234 0039d316-1c4b-4281-b951-d872f2087c98