File Manager: refactor metadata readers, show metadata in UI
This is a work-in-progress refactoring.  I was hoping to have it done by now, but just discovered today that Chrome already has support for some of the ByteReader functionality, including the only reliable way to deal with signed/unsigned numbers.

This isn't working yet, and so it may change a bit more before landing, but please feel free to suggest changes now.

BUG=chromium-os:17498
TEST=manual testing on chromeos-chrome on linux

Review URL: http://codereview.chromium.org/7322001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@92201 0039d316-1c4b-4281-b951-d872f2087c98