Revert 91769 - Revert 89600 - Treat ERR_CONNECTION_CLOSED as a succesful download.
Some servers will send a Content-Length which exceeds the number of bytes sent
in the message body.

Although this violates the RFC spec, other major browsers permit it
and treat a download as complete in that case. Chrome did as well
prior to M12, and now some users are unable to access the item that
they downloaded.

We ran an experiment that caused an error in this case, to gather UMA statistics.  This reverts that experiment.

BUG=85549
TEST=Go to server which advertises larger Content-Length than number of bytes
sent, confirm that the download is again reported as interrupted.


Review URL: http://codereview.chromium.org/7353023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@92697 0039d316-1c4b-4281-b951-d872f2087c98