NaCl: Enable tests on Mac OS X
The tests were partly enabled on Mac before, but with the return code
ignored.  Now that we've seen the tests mostly work on the bots,
enable them properly, but with two disabled.

Update the comment to explain how to disable tests.

Remove the disabling of "ppapi_ppb_scrollbar_browser_test", which I
did erroneously: The test started passing in the change where I added
this (because I updated DEPS to a revision where it passed), and
anyway I got the test name wrong (omitting the "run_" suffix).

Add a flush() call because otherwise the log message does not appear
until after the tests on Windows.

BUG=none
TEST=nacl_integration

Review URL: http://codereview.chromium.org/7484046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@93719 0039d316-1c4b-4281-b951-d872f2087c98