Reland 87645 with clang fixes.
[Sync] Refactor sync datatype error handling.

This introduces SyncError's, which are a convenient way of passing around an
error location, type, and message. All datatypes have been refactored to use
this, including in the AssociateModels code. A future change will use this
to add support for continuing sync even when a datatype fails to start.
In addition, eventually a future change will convert the UnrecoverableError
handler to use SyncError's as well as have the datatype controller's and
datatype manager surface SyncError's to the PSS.

Original Review URL: http://codereview.chromium.org/7453014

TBR=akalin@chromium.org
BUG=87645
TEST=

Review URL: http://codereview.chromium.org/7461098

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@94204 0039d316-1c4b-4281-b951-d872f2087c98