Sync performance tests: Cleanup TODOs and print in dashboard friendly format
This patch adds a new PrintResult method in SyncTimingHelper that should print timings in a format appropriate for use with the chromium perf dashboard.  It is a much simpler replication of the methods found in chrome/test/ui/ui_perf_test.{h,cc}.

Also cleanup TODOs that are no longer needed or relevant.

BUG=90441
TEST=sync_performance_tests

Review URL: http://codereview.chromium.org/7464046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@94224 0039d316-1c4b-4281-b951-d872f2087c98