Allow sync integration tests to operate on multiple datatypes: Preferences + Bookmarks
The sync integration tests currently use a class hierarchy where the test classes for each datatype are subclasses of LiveSyncTest. While this design worked in the past, it allows tests to work with only one datatype at a time, and therefore doesn't allow us to test the interplay between datatypes.

This patch is the first in a series of patches that will move away from an inheritance model to one where test cases can operate on more than one datatype. It updates the Preferences and Bookmarks datatypes to the new model, and contains the following changes:

- LivePreferencesSyncTest no longer inherits from LiveSyncTest, but is renamed to PreferencesHelper, and contains a bunch of static methods that perform various operations related to preferences. Similarly, LiveBookmarksSyncTest becomes BookmarksHelper.
- SingleClientLivePreferencesSyncTest is renamed to SingleClientPreferencesSyncTest, and is a subclass of LiveSyncTest. It uses the methods in PreferencesHelper by including its header file and attaching itself to the helper. Similarly, SingleClientBookmarksSyncTest uses the methods in BookmarksHelper.
- The old helper class BookmarkModelVerifier is now refactored into BookmarksHelper. This eliminates a bunch of wrapper methods that were in LiveBookmarksSyncTest.
- Once more datatype helper classes are refactored out, tests can attach to more than one datatype helper and use all their methods. For example, the class MigrationErrorsTest benefits from this model by being able to use PreferencesHelper and BookmarksHelper.

BUG=88510
TEST=sync_integration_tests, sync_performance_tests

Review URL: http://codereview.chromium.org/7259005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@94437 0039d316-1c4b-4281-b951-d872f2087c98