Fix bad section object leak when painting plugins
The skia canvas does not own the section handle in: PlatformCanvas::initialize(x, y, true, section)
so you have close it at some point.

Since skia calls CreateDibSection in this fuction, which maps the memory, it is safe to close the
handle after initialize. A test is being added to ui_tests to make sure we know if this works.


BUG=31173
TEST= see http://codereview.chromium.org/7492074/
Review URL: http://codereview.chromium.org/7482047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@94604 0039d316-1c4b-4281-b951-d872f2087c98