Sync Promo: Show profile avatar next to user name
If the user is signed into sync then we now show the profile avatar next to their user name in the sync promo. We only show the profile avatar icon if the user has created multiple profiles (this matches the behavior of the avatar button in the browser window).

This change also adds hooks to show the profile menu if the user clicks their user name. Currently this doesn't do anything. I'll change it to show the profile menu once the profile menu CL lands.

BUG=91009
TEST=


Review URL: http://codereview.chromium.org/7610008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@96527 0039d316-1c4b-4281-b951-d872f2087c98